pyproject.toml
[project]
name = "fisherman"
version = "0.1.0"
description = "Lightweight macOS screen streamer"
requires-python = ">=3.12"
dependencies = [
    "pyobjc-core>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-Vision>=10.0",
    "Pillow>=10.0",
    "imagehash>=4.3",
    "websockets>=13.0",
    "pydantic-settings>=2.0",
    "click>=8.0",
    "structlog>=24.0",
]

[project.scripts]
fisherman = "fisherman.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
